fix: 下書きルートで記事が見つからない問題を修正
