+++ src/app/articles/[articleId]/page.tsx
@@ -1,7 +1,19 @@
 import { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 import ArticleContent from '@/components/ArticleContent';
-import { getAllArticleIds, getArticle } from '@/utils/micro-cms';
+import LoadingSkeleton from '@/components/LoadingSkeleton';
+import MemberInfo from '@/components/MemberInfo';
+import TwoColumnLayout from '@/components/layout/TwoColumnLayout';
+
+import {
+  getAllArticleIds,
+  getArticle,
+  getOtherArticlesBySameMember,
+  getRecommendedArticles,
+} from '@/utils/micro-cms';
+import ArticleCard from '@/components/ArticleCard';
+import HeadingH2 from '@/components/HeadingH2';
+import { Suspense } from 'react';
 
 type Params = {
   params: { articleId: string };
@@ -34,10 +46,64 @@ export async function generateMetadata({
   }
 }
 
+// これはdraftルートで使ってはいけない。
+// `draftKey`無しのレスポンスがキャッシュされてしまう
+
+/** これが完了する前にページを表示できるように分離 */
+async function OtherArticles({ articleId }: { articleId: string }) {
+  const otherArticles = await getOtherArticlesBySameMember(articleId);
+  return otherArticles && otherArticles.contents.length > 0 ? (
+    <div className="pt-24 text-center sm:mx-10">
+      <>
+        <HeadingH2>この人が書いた記事</HeadingH2>
+        {otherArticles.contents.map((article, index) => (
+          <ArticleCard key={index} article={article} className="py-8" />
+        ))}
+      </>
+    </div>
+  ) : null;
+}
+
+/** これが完了する前にページを表示できるように分離 */
+async function RecommendedArticles({ articleId }: { articleId: string }) {
+  const recommendedArticles = await getRecommendedArticles(articleId);
+  {
+    /* <!-- ▼ 最新のオススメ記事 --> */
+  }
+  return recommendedArticles && recommendedArticles.contents.length > 0 ? (
+    <div className="pt-24 text-center sm:mx-10">
+      <HeadingH2>最新のオススメ記事</HeadingH2>
+      {recommendedArticles.contents.map((article, index) => (
+        <ArticleCard key={index} article={article} className="py-8" />
+      ))}
+    </div>
+  ) : null;
+}
+
 export default async function ArticlePage({ params: { articleId } }: Params) {
   const article = await getArticle(articleId);
   if (!article) {
     notFound();
   }
-  return <ArticleContent article={article} />;
+  return (
+    <TwoColumnLayout
+      sidebarChildren={
+        <>
+          <MemberInfo member={article.name} />
+          {/* <!-- ▼ 同部員の他の記事 --> */}
+          <Suspense fallback={<LoadingSkeleton />}>
+            {/* @ts-expect-error Server Component */}
+            <OtherArticles articleId={articleId} />
+          </Suspense>
+          {/* <!-- ▼ 最新のオススメ記事 --> */}
+          <Suspense fallback={<LoadingSkeleton />}>
+            {/* @ts-expect-error Server Component */}
+            <RecommendedArticles articleId={articleId} />
+          </Suspense>
+        </>
+      }
+    >
+      <ArticleContent article={article} />
+    </TwoColumnLayout>
+  );
 }
